+++ packages/decap-cms-backend-azure/src/AuthenticationPage.js
@@ -23,6 +23,14 @@ export default class AzureAuthenticationPage extends React.Component {
   state = {};
 
   componentDidMount() {
+    // Manually validate PropTypes - React 19 breaking change
+    PropTypes.checkPropTypes(
+      AzureAuthenticationPage.propTypes,
+      this.props,
+      'prop',
+      'AzureAuthenticationPage',
+    );
+
     this.auth = new ImplicitAuthenticator({
       base_url: `https://login.microsoftonline.com/${this.props.config.backend.tenant_id}`,
       auth_endpoint: 'oauth2/authorize',
@@ -65,7 +73,8 @@ export default class AzureAuthenticationPage extends React.Component {
         onLogin={this.handleLogin}
         loginDisabled={inProgress}
         loginErrorMessage={this.state.loginError}
-        logoUrl={config.logo_url}
+        logoUrl={config.logo_url} // Deprecated, replaced by `logo.src`
+        logo={config.logo}
         renderButtonContent={() => (
           <React.Fragment>
             <LoginButtonIcon type="azure" />