+++ packages/decap-cms-backend-gitea/src/AuthenticationPage.js
@@ -19,6 +19,14 @@ export default class GiteaAuthenticationPage extends React.Component {
   state = {};
 
   componentDidMount() {
+    // Manually validate PropTypes - React 19 breaking change
+    PropTypes.checkPropTypes(
+      GiteaAuthenticationPage.propTypes,
+      this.props,
+      'prop',
+      'GiteaAuthenticationPage',
+    );
+
     const { base_url = 'https://try.gitea.io', app_id = '' } = this.props.config.backend;
     this.auth = new PkceAuthenticator({
       base_url,
@@ -56,7 +64,8 @@ export default class GiteaAuthenticationPage extends React.Component {
         onLogin={this.handleLogin}
         loginDisabled={inProgress}
         loginErrorMessage={this.state.loginError}
-        logoUrl={config.logo_url}
+        logoUrl={config.logo_url} // Deprecated, replaced by `logo.src`
+        logo={config.logo}
         siteUrl={config.site_url}
         renderButtonContent={() => (
           <React.Fragment>