+++ packages/decap-cms-backend-proxy/src/AuthenticationPage.js
@@ -36,6 +36,16 @@ export default class AuthenticationPage extends React.Component {
     t: PropTypes.func.isRequired,
   };
 
+  componentDidMount() {
+    // Manually validate PropTypes - React 19 breaking change
+    PropTypes.checkPropTypes(
+      AuthenticationPage.propTypes,
+      this.props,
+      'prop',
+      'AuthenticationPage',
+    );
+  }
+
   handleLogin = e => {
     e.preventDefault();
     this.props.onLogin(this.state);
@@ -46,7 +56,7 @@ export default class AuthenticationPage extends React.Component {
 
     return (
       <StyledAuthenticationPage>
-        {renderPageLogo(config.logo_url)}
+        {renderPageLogo(config.logo?.src || config.logo_url)}
         <LoginButton disabled={inProgress} onClick={this.handleLogin}>
           {inProgress ? t('auth.loggingIn') : t('auth.login')}
         </LoginButton>