+++ packages/decap-cms-core/index.d.ts
@@ -140,6 +140,8 @@ declare module 'decap-cms-core' {
     default?: any;
 
     allow_add?: boolean;
+    allow_remove?: boolean;
+    allow_reorder?: boolean;
     collapsed?: boolean;
     summary?: string;
     minimize_collapsed?: boolean;
@@ -240,6 +242,7 @@ declare module 'decap-cms-core' {
     // This is the default widget, so declaring its type is optional.
     widget?: 'string' | 'text';
     default?: string;
+    visualEditing?: boolean;
   }
 
   export interface CmsFieldMeta {
@@ -313,6 +316,7 @@ declare module 'decap-cms-core' {
     hide?: boolean;
     editor?: {
       preview?: boolean;
+      visualEditing?: boolean;
     };
     publish?: boolean;
     nested?: {
@@ -389,7 +393,11 @@ declare module 'decap-cms-core' {
     locale?: string;
     site_url?: string;
     display_url?: string;
-    logo_url?: string;
+    logo_url?: string; // Deprecated, replaced by `logo.src`
+    logo?: {
+      src: string;
+      show_in_header?: boolean;
+    };
     show_preview_links?: boolean;
     media_folder?: string;
     public_folder?: string;