+++ packages/decap-cms-core/src/types/redux.ts
@@ -155,6 +155,8 @@ export interface CmsFieldList {
   default?: unknown;
 
   allow_add?: boolean;
+  allow_remove?: boolean;
+  allow_reorder?: boolean;
   collapsed?: boolean;
   summary?: string;
   minimize_collapsed?: boolean;
@@ -255,6 +257,7 @@ export interface CmsFieldStringOrText {
   // This is the default widget, so declaring its type is optional.
   widget?: 'string' | 'text';
   default?: string;
+  visualEditing?: boolean;
 }
 
 export interface CmsFieldMeta {
@@ -329,6 +332,7 @@ export interface CmsCollection {
   delete?: boolean;
   editor?: {
     preview?: boolean;
+    visualEditing?: boolean;
   };
   publish?: boolean;
   nested?: {
@@ -402,7 +406,11 @@ export interface CmsConfig {
   locale?: string;
   site_url?: string;
   display_url?: string;
-  logo_url?: string;
+  logo_url?: string; // Deprecated, replaced by `logo.src`
+  logo?: {
+    src: string;
+    show_in_header?: boolean;
+  };
   show_preview_links?: boolean;
   media_folder?: string;
   public_folder?: string;
@@ -534,6 +542,8 @@ export type EntryObject = {
   slug: string;
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   data: any;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  i18n?: any;
   collection: string;
   mediaFiles: List<MediaFileMap>;
   newRecord: boolean;