+++ packages/decap-cms-ui-auth/src/NetlifyAuthenticationPage.js
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import styled from '@emotion/styled';
-import { partial } from 'lodash';
+import partial from 'lodash/partial';
 import {
   AuthenticationPage,
   buttons,
@@ -65,8 +65,16 @@ if (window.netlifyIdentity) {
   });
 }
 
-export default class GitGatewayAuthenticationPage extends React.Component {
+export default class NetlifyAuthenticationPage extends React.Component {
   static authClient;
+
+  static propTypes = {
+    onLogin: PropTypes.func.isRequired,
+    inProgress: PropTypes.bool.isRequired,
+    error: PropTypes.node,
+    config: PropTypes.object.isRequired,
+    t: PropTypes.func.isRequired,
+  };
 
   constructor(props) {
     super(props);
@@ -74,6 +82,14 @@ export default class GitGatewayAuthenticationPage extends React.Component {
   }
 
   componentDidMount() {
+    // Manually validate PropTypes - React 19 breaking change
+    PropTypes.checkPropTypes(
+      NetlifyAuthenticationPage.propTypes,
+      this.props,
+      'prop',
+      'GitGatewayAuthenticationPage',
+    );
+
     if (!this.loggedIn && window.netlifyIdentity && window.netlifyIdentity.currentUser()) {
       this.props.onLogin(window.netlifyIdentity.currentUser());
       window.netlifyIdentity.close();
@@ -111,14 +127,6 @@ export default class GitGatewayAuthenticationPage extends React.Component {
     }
   };
 
-  static propTypes = {
-    onLogin: PropTypes.func.isRequired,
-    inProgress: PropTypes.bool.isRequired,
-    error: PropTypes.node,
-    config: PropTypes.object.isRequired,
-    t: PropTypes.func.isRequired,
-  };
-
   state = { email: '', password: '', errors: {} };
 
   handleChange = (name, e) => {
@@ -144,7 +152,7 @@ export default class GitGatewayAuthenticationPage extends React.Component {
     }
 
     try {
-      const client = await GitGatewayAuthenticationPage.authClient();
+      const client = await NetlifyAuthenticationPage.authClient();
       const user = await client.login(this.state.email, this.state.password, true);
       this.props.onLogin(user);
     } catch (error) {
@@ -163,7 +171,8 @@ export default class GitGatewayAuthenticationPage extends React.Component {
       if (errors.identity) {
         return (
           <AuthenticationPage
-            logoUrl={config.logo_url}
+            logoUrl={config.logo_url} // Deprecated, replaced by `logo.src`
+            logo={config.logo}
             siteUrl={config.site_url}
             onLogin={this.handleIdentity}
             renderPageContent={() => (
@@ -181,7 +190,8 @@ export default class GitGatewayAuthenticationPage extends React.Component {
       } else {
         return (
           <AuthenticationPage
-            logoUrl={config.logo_url}
+            logoUrl={config.logo_url} // Deprecated, replaced by `logo.src`
+            logo={config.logo}
             siteUrl={config.site_url}
             onLogin={this.handleIdentity}
             renderButtonContent={() => t('auth.loginWithNetlifyIdentity')}
@@ -193,7 +203,8 @@ export default class GitGatewayAuthenticationPage extends React.Component {
 
     return (
       <AuthenticationPage
-        logoUrl={config.logo_url}
+        logoUrl={config.logo_url} // Deprecated, replaced by `logo.src`
+        logo={config.logo}
         siteUrl={config.site_url}
         renderPageContent={() => (
           <AuthForm onSubmit={this.handleLogin}>