+++ packages/decap-cms-ui-default/src/Icon/images/_index.js
@@ -42,6 +42,7 @@ import iconRefresh from './refresh.svg';
 import iconScroll from './scroll.svg';
 import iconSearch from './search.svg';
 import iconSettings from './settings.svg';
+import iconStrikethrough from './strikethrough.svg';
 import iconUser from './user.svg';
 import iconWorkflow from './workflow.svg';
 import iconWrite from './write.svg';
@@ -94,6 +95,7 @@ const images = {
   scroll: iconScroll,
   search: iconSearch,
   settings: iconSettings,
+  strikethrough: iconStrikethrough,
   user: iconUser,
   workflow: iconWorkflow,
   write: iconWrite,