+++ packages/decap-cms-widget-list/src/ListControl.js
@@ -4,7 +4,9 @@ import ImmutablePropTypes from 'react-immutable-proptypes';
 import styled from '@emotion/styled';
 import { css, ClassNames } from '@emotion/react';
 import { List, Map, fromJS } from 'immutable';
-import { partial, isEmpty, uniqueId } from 'lodash';
+import partial from 'lodash/partial';
+import isEmpty from 'lodash/isEmpty';
+import uniqueId from 'lodash/uniqueId';
 import { v4 as uuid } from 'uuid';
 import DecapCmsWidgetObject from 'decap-cms-widget-object';
 import {
@@ -217,6 +219,11 @@ export default class ListControl extends React.Component {
       value: this.valueToString(value),
       keys,
     };
+  }
+
+  componentDidMount() {
+    // Manually validate PropTypes - React 19 breaking change
+    PropTypes.checkPropTypes(ListControl.propTypes, this.props, 'prop', 'ListControl');
   }
 
   valueToString = value => {
@@ -675,6 +682,8 @@ export default class ListControl extends React.Component {
           onCollapseToggle={partial(this.handleItemCollapseToggle, index)}
           dragHandle={DragHandle}
           id={key}
+          allowRemove={field.get('allow_remove', true)}
+          allowReorder={field.get('allow_reorder', true)}
           onRemove={partial(this.handleRemove, index)}
           data-testid={`styled-list-item-top-bar-${key}`}
         />