+++ packages/decap-cms-widget-markdown/src/MarkdownControl/VisualEditor.js
@@ -9,7 +9,7 @@ import { createEditor, Transforms, Editor as SlateEditor } from 'slate';
 import { Editable, ReactEditor, Slate, withReact } from 'slate-react';
 import { withHistory } from 'slate-history';
 import { fromJS } from 'immutable';
-import { isEqual } from 'lodash';
+import isEqual from 'lodash/isEqual';
 
 import { editorStyleVars, EditorControlBar } from '../styles';
 import Toolbar from './Toolbar';
@@ -141,7 +141,8 @@ function Editor(props) {
 
   function handleMarkClick(format) {
     ReactEditor.focus(editor);
-    toggleMark(editor, format);
+    const markFormat = format === 'strikethrough' ? 'delete' : format;
+    toggleMark(editor, markFormat);
   }
 
   function handleBlockClick(format) {
@@ -222,7 +223,7 @@ function Editor(props) {
         position: relative;
       `}
     >
-      <Slate editor={editor} value={editorValue} onChange={handleChange}>
+      <Slate editor={editor} value={editorValue} initialValue={editorValue} onChange={handleChange}>
         <EditorControlBar>
           {
             <Toolbar