+++ packages/decap-cms-lib-util/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.3.1](https://github.com/decaporg/decap-cms/compare/decap-cms-lib-util@3.3.0...decap-cms-lib-util@3.3.1) (2025-07-31)
+
+**Note:** Version bump only for package decap-cms-lib-util
+
+# [3.3.0](https://github.com/decaporg/decap-cms/compare/decap-cms-lib-util@3.2.0...decap-cms-lib-util@3.3.0) (2025-06-26)
+
+**Note:** Version bump only for package decap-cms-lib-util
+
 # [3.2.0](https://github.com/decaporg/decap-cms/compare/decap-cms-lib-util@3.1.0...decap-cms-lib-util@3.2.0) (2025-01-29)
 
 ### Features