+++ packages/decap-cms-widget-datetime/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.4.0](https://github.com/decaporg/decap-cms/compare/decap-cms-widget-datetime@3.2.3...decap-cms-widget-datetime@3.4.0) (2025-07-15)
+
+**Note:** Version bump only for package decap-cms-widget-datetime
+
+# [3.3.0](https://github.com/decaporg/decap-cms/compare/decap-cms-widget-datetime@3.2.3...decap-cms-widget-datetime@3.3.0) (2025-06-26)
+
+**Note:** Version bump only for package decap-cms-widget-datetime
+
 ## [3.2.3](https://github.com/decaporg/decap-cms/compare/decap-cms-widget-datetime@3.2.2...decap-cms-widget-datetime@3.2.3) (2024-08-30)
 
 ### Bug Fixes